refactor(snapshot): rename Lock to Snapshot across project
