+++ src/Project/Snapshot/Snapshot.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Project\Lock;
+namespace Haspadar\Piqule\Project\Snapshot;
 
 use Haspadar\Piqule\Target\Target;
 
-interface Lock
+interface Snapshot
 {
     public function has(Target $target): bool;
 
     public function hashOf(Target $target): string;
 
-    public function with(Target $target): Lock;
+    public function with(Target $target): Snapshot;
 
     public function store(): void;
 }
