+++ src/Project/UninitializedProject.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Haspadar\Piqule\Project;
 
 use Haspadar\Piqule\PiquleException;
-use Haspadar\Piqule\Project\Lock\Lock;
+use Haspadar\Piqule\Project\Snapshot\Snapshot;
 use Haspadar\Piqule\Source\SourceDirectory;
 use Haspadar\Piqule\Target\DiskTarget;
 use Haspadar\Piqule\Target\Materialization\Materialization;
@@ -18,19 +18,19 @@ final readonly class UninitializedProject implements Project
         private TargetDirectory $targetDirectory,
     ) {}
 
-    public function init(Materialization $materialization, Lock $lock): void
+    public function init(Materialization $materialization, Snapshot $snapshot): void
     {
         foreach ($this->sourceDirectory->files() as $sourceFile) {
-            $lock = $materialization->applyTo(
+            $snapshot = $materialization->applyTo(
                 new DiskTarget($sourceFile, $this->targetDirectory),
-                $lock,
+                $snapshot,
             );
         }
 
-        $lock->store();
+        $snapshot->store();
     }
 
-    public function update(Materialization $materialization, Lock $lock): void
+    public function update(Materialization $materialization, Snapshot $snapshot): void
     {
         throw new PiquleException('Project is not initialized');
     }
